Start on dbaccess and backend
Created dbaccess with helper function executeSQL() as a streamlined way to access the database.
Added code to create text file storing path for the database.
Added subroutines in backend to add, search, edit and remove both instructors and customers. subroutines for login and change password and get permissions for a certain account. added subroutine to add certification columns to the instructor table.